Arbiter fix
* Fixed sending more troops than available

@NoctuaNivalis is a dirty cheater

* Fixed unrightfull draws

When multiple marches with the same source and destination where issues in the same turn,
the marches got merged by adding the troops together. However, the second (and third, fourth, ...),
were not killed off. This meant that the player had a reference to a march that could never die (it wasn't
on any road, so not encoutering any other troops), failing a wincondition for the other player (a player can't lose
if he has any more marches).